Renamed/reorganized directories. Deleted Backpack component
