+++ src/engine/components/CharacterController.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "engine/Component.hpp"
+#include "core/Component.hpp"
 #include "physics/CharacterPhysicsUpdate.hpp"
 
 namespace lei3d
