+++ src/engine/components/SkyBox.cpp
@@ -1,6 +1,6 @@
 #include "SkyBox.hpp"
 
-#include "util/GLDebug.hpp"
+#include "logging/GLDebug.hpp"
 
 namespace lei3d
 {
