+++ src/engine/components/StaticCollider.cpp
@@ -1,5 +1,5 @@
 #include "StaticCollider.hpp"
-#include "util/Log.hpp"
+#include "logging/Log.hpp"
 
 namespace lei3d
 {
