+++ src/engine/core/Component.cpp
@@ -1,6 +1,6 @@
 #include "Component.hpp"
 
-#include "engine/Application.hpp"
+#include "core/Application.hpp"
 
 namespace lei3d {
 	Component::Component(Entity* entity) : m_Entity(entity) {
