+++ src/engine/rendering/Mesh.cpp
@@ -1,6 +1,6 @@
 #include "Mesh.hpp"
 
-#include "util/GLDebug.hpp"
+#include "logging/GLDebug.hpp"
 
 namespace lei3d
 {
