+++ src/engine/rendering/Shader.cpp
@@ -1,5 +1,5 @@
 ﻿#include "Shader.hpp"
-#include "util/GLDebug.hpp"
+#include "logging/GLDebug.hpp"
 
 namespace lei3d 
 {
