+++ src/engine/scenes/Scene.cpp
@@ -1,6 +1,6 @@
 #include "scenes/Scene.hpp"
 
-#include "util/GLDebug.hpp"
+#include "logging/GLDebug.hpp"
 
 namespace lei3d {
 	Scene::Scene() {
